Load Beta Inside PDF (#26)
create a base64 encoded string with open ssl and drop it into the file.

The architecture is super hacky still. It would be nice to automate the base64 string creation process and add the string to the assets and then load it into the redux store as the intial state.

But I don't have time to make it right, so I've just made it work. We'll circle back on this after the event